Fixed the admin vs user problem in the api routes. Also added a middleware file in API routes for reuseable middleware in routes.
